requirements.txt
Adafruit-Blinka==5.4.0
adafruit-circuitpython-busdevice==5.0.1
adafruit-circuitpython-mcp3xxx==1.4.2
Adafruit-PlatformDetect==2.17.0
Adafruit-PureIO==1.1.5
atomicwrites==1.4.0
attrs==20.1.0
board==1.0
boto3==1.16.56
botocore==1.19.56
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.3
DateTime==4.3
Flask==1.1.2
Flask-GoogleCharts==0.1.0
Flask-Injector==0.12.3
idna==2.10
iniconfig==1.0.1
injector==0.18.3
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
MarkupSafe==1.1.1
more-itertools==8.5.0
oauthlib==3.1.0
packaging==20.4
pluggy==0.13.1
py==1.9.0
pyftdi==0.51.2
pyparsing==2.4.7
pyserial==3.4
pytest==6.0.1
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
pyusb==1.1.0
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
s3transfer==0.3.4
six==1.15.0
toml==0.10.1
twitter==1.18.0
typing-extensions==3.7.4.3
urllib3==1.26.2
Werkzeug==1.0.1
zope.interface==5.1.0
